Create check_grin_wallet.sh
The script will check grin wallet online or not associate with the IP addresss that the wallet is running. Then the script sends email to you for latest status, we can use crontab to check the grin wallet periodically.
Require to have input file grin_wallet_addresses_with_ips.txt with content "grin_wallet_adress(space)IP_address"
In case the server can't send mail, install ISPConfig or postfix to relay mail to a trusted SMTP to send it to you.